+++ server/middleware/s3.ts
@@ -1,3 +1,5 @@
+import { join } from "path";
+
 // Middleware to intercept requests and proxy if certain conditions are met
 const config = {
   endpoint: process.env.AWS_ENDPOINT,
@@ -11,37 +13,45 @@ const config = {
 const S3_BUCKET_NAME = process.env.S3_BUCKET_NAME;
 
 const fs = require("@cyclic.sh/s3fs")(S3_BUCKET_NAME, config);
-import { join } from "path";
-import { getJanDataFolderPath, normalizeFilePath } from "@janhq/core/node";
+const PROXY_PREFIX = "/v1/fs";
+const PROXY_ROUTES = ["/threads", "/messages"];
 
 export const s3 = (req: any, reply: any, done: any) => {
   // Proxy FS requests to S3 using S3FS
-  if (req.url.startsWith("/v1/fs")) {
+  if (req.url.startsWith(PROXY_PREFIX)) {
     const route = req.url.split("/").pop();
     const args = parseRequestArgs(req);
-    try {
-      if (args[0].includes("/threads") || args[0].includes("/messages")) {
+
+    // Proxy matched requests to the s3fs module
+    if (args.length && PROXY_ROUTES.some((route) => args[0].includes(route))) {
+      try {
         // Handle customized route
+        // S3FS does not handle appendFileSync
         if (route === "appendFileSync") {
           let result = handAppendFileSync(args);
 
           reply.status(200).send(result);
           return;
         }
-        // Proxy the request to the s3fs module
+        // Reroute the other requests to the s3fs module
         const result = fs[route](...args);
         reply.status(200).send(result);
         return;
+      } catch (ex) {
+        console.log(ex);
       }
-    } catch (ex) {
-      console.log(ex);
     }
   }
-  // Let the request go through for other routes
+  // Let other requests go through
   done();
 };
 
 const parseRequestArgs = (req: Request) => {
+  const {
+    getJanDataFolderPath,
+    normalizeFilePath,
+  } = require("@janhq/core/node");
+
   return JSON.parse(req.body as any).map((arg: any) =>
     typeof arg === "string" &&
     (arg.startsWith(`file:/`) || arg.startsWith(`file:\\`))
